requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask[async]==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==3.0.1
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
style==1.1.0
tzdata==2024.2
update==0.0.1
urllib3==2.2.3
Werkzeug==3.0.4
asgiref==3.7.2
redis==5.0.1
aioredis==2.0.1 
jsonschema==4.20.0